JDCP-7: Clean up sockets on disconnect/connect race.
Motivation
----------
If someone disconnects very quickly, even before the connect
phase finished the channel is not set and as a result not
cleaned up and the file descriptor is leaked.

Modifications
-------------
The code now keeps track for all explicit disconnect operations
and once a connect finishes checks if set. If it is, then it will
proactively disconnect that socket in the background to avoid
leaking FD's.

Result
------
No leaking FDs in case connect and disconnect race each other.