JDCP-42: Do not fail if SessionState don't initialize all partitions
Motivation
----------
The caller might leave some partitions uninitilized, and if
it does not request to open stream for them, it should be
totally safe.

Modification
------------
Instead of just checking first initialized partition, check
if all requested partitions are initialized, and also check
total number of requested partitions do not exceeds the total
number of partitions of the cluster.

Result
------
More flexible and informative sanity check