Add benchmark for streaming from Couchbase server.
POM changes are based on https://stackoverflow.com/a/39374972

The added benchmarks require a server running on localhost, populated with "travel-sample" data.

To run the benchmarks:

    $ mvn clean package -DskipTests
    $ java -jar target/dcp-client-<VERSION>-perf-tests.jar
           (add -h to see list of JMH command line options)

The "ConnectAndStream" benchmark measures the time it takes to connect to the Couchbase server,
stream all the data from the "travel-sample" bucket, and disconnect.

The "StreamBenchmark" is a work in progress. It attempts to improve on "ConnectAndStream" by
reusing the same Client for each invocation in order to exclude the time it takes to connect to the server.
It's currently broken though; perhaps it's missing some important step to reset the client properly
between streaming sessions.

Command to run just the working benchmark:

    $ java -jar target/dcp-client-<VERSION>-perf-tests.jar ConnectAndStream