JDCP-52: Implement broken connection detection
If the "enable_noop" control is enabled, the client will restart a
connection when the server has not sent any messages for a period of
time greater than twice the noop interval.

The noop interval is specified by the client with the
"set_noop_interval" control. If this control is not explicitly set by
the application code, the client uses a default value of 120 seconds,
overriding the server-side default. This is done so the client always
knows the interval value.

Change-Id: I4f39648c0cc7aab9e5fea486b5efb3fe5455f70b
Reviewed-on: http://review.couchbase.org/82282
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: David Nault <david.nault@couchbase.com>