Remove work-in-progress integration tests
Motivation
----------
Groundwork for adopting the maven-failsafe-plugin.

Unlike the previous scheme of conditional test compilation based on
a profile, the failsafe plugin always compiles the integration tests.
This is a problem for the Jenkins job, since it uses JDK 1.6 but
the integration tests require 1.8. So for now we'll remove the
integration tests (which weren't really testing anything yet).
We can restore the integration test from git history later on

Modifications
-------------
Delete the integration test and support classes.

Result
------
maven-failsafe-plugin can be added in a future commit.

Integration test are gone, but can be retrieved from git history
when we want to put more effort into making them actually test stuff.

Change-Id: I18a034ee5d79782e9861324c0a954835b943ee8f
Reviewed-on: http://review.couchbase.org/82933
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Matt Ingenthron <matt@couchbase.com>