JDCP-66 Spurious stream reconnection when current seqno > end seqno
Motivation
==========
When streaming from BEGINNING to NOW, after closing a partition
the stream might get reopened due to the sequence number exceeding
the requests end seqno. The reopened stream then will trigger a series
of RollbackExceptions and eventually a loop from which the client might
not be able to recover.

Modifications
=============
SessionState.isAtEnd() and PartitionState.isAtEnd() are modified to
use >= instead of == when comparing the current seqno to the requested
end seqno.

Result
======
The isAtEnd() methods now return "true" even if the server has sent
more events than the user requested.

Conductor.maybeMovePartition() no longer spuriously reopens
the stream if the current seqno exceeds the end seqno.

Change-Id: Ifdf87975171cc7c35b9786a5c999c2f09b106235
Reviewed-on: http://review.couchbase.org/87509
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Build Bot <build@couchbase.com>