JDCP-73 HelloRequest.init() fails to set extras after first time
Motivation
----------
If HelloRequest.init() is called more than once, the extras are
not set on the subsequent requests because the buffer used to
populate the extras is shared, and its readable bytes get consumed.

Modifications
-------------
Set the extras using a slice of the shared buffer.

Result
------
Each call to init() uses a separate slice with its own reader index.
Subsequent calls set the extras correctly.

Change-Id: Ia913f8601ffc962ad3003373d2aae6cda8f74130
Reviewed-on: http://review.couchbase.org/89299
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>