JDCP-74 Enhanced compression support (followup)
Simplify the compression/decompression mode names:

    CompressionMode.DISCRETIONARY -> ENABLED

    DecompressionMode.TRANSPARENT -> ENABLED
    DecompressionMode.MANUAL -> DISABLED

Change the default compression mode to DISABLED, because the
performance impact of enabling compression has not yet been measured.

If the user asks for CompressionMode.ENABLED with a pre-Vulcan server,
treat it the same as if they had specified CompressionMode.FORCED.
The rationale is that if the user says they want compression, then
we should give them compression. This behavior may change in the future
if we decide to make ENABLED the default mode.

Change-Id: I01acb98c05936af9e8af450bbedf43ce3b03a877
Reviewed-on: http://review.couchbase.org/89435
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>