JDCP-78 Don't send buffer ack for "Close Stream" response
Motivation
==========
The client sends a buffer ack for the "close stream" response,
but this is not one of the messages for which the producer expects
an acknowledgement.

See https://github.com/couchbaselabs/dcp-documentation/blob/master/documentation/flow-control.md#buffering-messages

The consequences of acking too much are perhaps not as severe as not
acking enough (in fact it may even be benign, though I'm not sure),
but removing the incorrect ACK from filterDcpCloseStreamResponse()
simplifies future work.

Modifications
=============
* Don't ack CloseStream responses.
* Rename DcpStateVbucketStateMessage -> DcpSetVbucketStateMessage

Result
======
Client acks only the messages documented as requiring acknowledgement.

Change-Id: Ic3276e61df0eb8503347c47d79a42d27b2a1eba9
Reviewed-on: http://review.couchbase.org/90460
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Dave Rigby <daver@couchbase.com>