JDCP-80 Rollback mitigation
Motivation
==========
Rollbacks are problematic for some users. For example, in the Kafka
connector, once an event is published to a topic there is no way
to retract it in the event of a rollback. As a result, the topic
may contain phantom events (from an alternate history) that no longer
exist in Couchbase.

Modification
============
Add a new `mitigateRollbacks` method to the client builder.
When this feature is enabled, events are not published to the
DataEventHandler and ControlEventHandler until persistence has been
observed on all nodes hosting the partition.

Add a new `flowControl` method to the client builder.
When rollback mitigation is active, flow control is also required,
since events may be buffered for an extended period of time.
The new method is shorthand for setting the CONNECTION_BUFFER
DCP control and the bufferAckWatermark.

Add `PersistencePollingHandler`. As soon as a channel becomes active,
this handler starts issuing repeated observeSeqnos requests for all
partitions hosted on the node, whether active or replica.

Add `PersistedSeqnos` to keep track of which seqnos have been persisted
on which nodes.

Add `StreamEventBuffer` to intercept and buffer control and data events.
The buffer's `onSeqnoPersisted` method is called periodically by the
`PersistencePollingHandlers`. Any buffered events with seqno <= the
persisted seqno are then propagated to the control and data handlers
set by the user.

Result
======
Library users who enabled rollback mitigation will not receive events
unless they have been persisted to disk everywhere. Rollbacks will
be transparent to the user, except in the unlikely scenario of total
data loss (rollback to zero).

Polling for persistence requires more network traffic and CPU usage.
Data events will take longer to reach the user's event handler.
These factors can be adjusted by tweaking the persistence polling
interval specified when rollback mitigation is enabled.

Change-Id: I97963ad878fa0f11af613786c99cfb517fd1cba6
Reviewed-on: http://review.couchbase.org/93307
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>