Add integration tests
This commit adds a new "integration-test" module and an "aggregator"
module that allows the integration-test module to use the DCP client
library as a dependency.

It's a separate module because it builds a Spring Boot webapp
that runs inside a Docker network alongside the Couchbase Server nodes.
This is a variation of the testing approach pioneered by the Analytics
team.

The aggregator project should arguably be the top-level project.
For now, it's deeply nested and out of the way so as not to interfere
with the upcoming merge of the rollback mitigation feature.

To run the integration tests, cd to `integration-test/aggregator`
and run `mvn verify`. A local Docker installation is required.

Change-Id: If69724af9e1ee20f1da62dc656cd523c7e634b3a
Reviewed-on: http://review.couchbase.org/92160
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Abdullah Amoudi <abdullah.alamoudi@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>