JDCP-84 Client fails to reestablish connections when streaming to infinity
Motivation
===========
Due to a regression introduced in 0.13.0, the client will not attempt
to reconnect when the stream ends unexpectedly, because it thinks
the desired end seqno has already been reached.

Modifications
=============
Change PartitionState.isAtEnd() so it never returns true if
we're streaming to infinity.

Result
======
Conductor.maybeMovePartition() now correctly attempts to reconnect
when streaming to infinity.

Change-Id: I326a15300bbe4e19f47c1501a1bcb24981cf4f82
Reviewed-on: http://review.couchbase.org/92746
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>