JDCP-85 Client stops receiving some events after failover
Motivation
==========
When a node with open streams is failed over, those streams should
be reopened on the node hosting the new active vbucket, but that
wasn't happening.

This surfaced as a problem in the Kafka connector, where users needed
to restart the connector after Couchbase Server node failover.

Modifications
=============
When removing a channel in response to a bucket config refresh,
call maybeMovePartition for each stream open on the doomed channel.
This causes the stream to be reopened on a less doomed channel.

When attempting to open a stream that is already open, allow the
attempt to succeed instead of reporting an error and incorrectly
marking the stream as closed. This mitigates a race condition where
a DCP channel's reconnect handler and the Conductor's config refresher
both call maybeMovePartition.

Result
======
Streaming resumes after failover.

Change-Id: Iab67fca63ea2035f76673ab8033b207bb1becc79
Reviewed-on: http://review.couchbase.org/95212
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>