JDCP-93 Rollback mitigation may produce obsolete data events
Motivation
==========
When rollback mitigation is enabled, if the connection drops
while there are unpersisted events in the stream event buffer,
the stream will be reopened starting from the last persisted event.
This is by design. However, the event buffer is not cleared,
so it contains obsolete entries from the previous stream.

In the worst case, this could result in events from a rolled-back
alternate history being propagated to the data event handler.

Modifications
=============
Whenever a stream is opened, clear the stream event buffer
for that vbucket. This is safe because the start seqno for the
"open stream" request is guaranteed to be lower than the lowest seqno
in the event buffer.

Add an integration test for this scenario.

Result
======
Unpersisted events from an alternate history are not passed to
the user's data handler.

Change-Id: I31afca35309d515f0e799a284170f2ef1c046757
Reviewed-on: http://review.couchbase.org/95648
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Build Bot <build@couchbase.com>