JDCP-94 Should never pass null ChannelFlowController to event handlers
Motivation
==========
NullPointerException observed in Kafka connector because its
control event handler attempts to acknowledge a rollback event.
It should be safe to do this, since the general contract is that
event handlers can attempt to ack anything, and it's up to the
flow controller to determine whether the event actually requires
acknowledgement.

Modifications
=============
Change ChannelFlowController to an interface.

Pass a dummy "do nothing" instance to the event handler instead of null.

Result
======
Event handlers can safely assume a non-null ChannelFlowController.

Change-Id: Ic3e20602f9fa5e2045af64abfb0a3e03ea52a7d8
Reviewed-on: http://review.couchbase.org/95711
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>