JDCP-95 Client fails to reopen stream if last message received was a snapshot marker
Motivation
==========
When opening a stream, the server complains if snapshotStartSeqno
is greater than startSeqno.

If the client receives a snapshot marker and then immediately
disconnects, the partition state will have a snapshotStartSeqno that
is greater than the last seen seqno. When attempting to reopen the
stream, the request fails because the arguments are invalid.

Modifications
=============
When opening the stream, first check whether the snapshotStartSeqno
is one greater than the startSeqno (as it would be in the scenario
described above). If so, disregard the snapshot marker and proceed
as if we had no snapshot info.

Result
======
Disconnecting immediately after receiving a snapshot marker no longer
causes subsequent openStream requests to fail.

Change-Id: I9a3677a8eed23cbb403e6df09bb86a5104906cce
Reviewed-on: http://review.couchbase.org/95980
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>