JDCP-98 Rollback mitigation: buffered events not released on rollback
Motivation
==========
When rollback mitigation is enabled, the ByteBufs associated with
buffered events are not released or ACK'd when the stream buffer is
cleared or rolled back.

This has the potential to exhaust the flow control buffer, and could
lead to memory leaks if the if the buffers are pooled.

Modifications
=============
When discarding a buffered event, attempt to ACK it (may fail if
channel is already closed) and always release the ByteBuf.

Fix typo: MAX_PARITITONS -> MAX_PARTITIONS

Result
======
Discarded events are ACK'd and released appropriately.

Change-Id: Id8d2acb83fc4706b80f359936882eeec7dac4ff6
Reviewed-on: http://review.couchbase.org/96473
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>