JDCP-109 Rollback mitigation: event buffer erroneously cleared in openStream response handler
Motivation
==========
When rollback mitigation is enabled and the DCP connection goes away,
any events remaining in the stream event buffer should be discarded.
The fix for JDCP-93 addressed this issue by clearing the stream buffer
when the openStream response is received. Unfortunately, due to an
apparent race condition, the openStream response handler can be invoked
*after* data events have already been received and added to the buffer.
Clearing the buffer at this point causes the early events to be lost.

Modifications
=============
Clear the event buffer when the channel is disconnected instead of
when a stream is opened.

Result
======
Buffer get cleared appropriately (integration test for this scenario
still passes), and the events that arrive shortly after the stream is
opened are no longer occasionally lost.

Change-Id: I069261100d033c2f7648a16ded3f19ce6ad16db9
Reviewed-on: http://review.couchbase.org/98408
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Abdullah Amoudi <abdullah.alamoudi@couchbase.com>