JDCP-116 BucketConfigHelper.getHostedPartitions() fails when node hosts no partitions
Motivation
==========
When persistence polling is enabled, the client was connecting to
all Couchbase nodes regardless of whether they hosted any partitions.
This triggered an exception in BucketConfigHelper.getHostedPartitions()
when called for those nodes that don't host any partitions.

Modifications
=============
Change NodeToPartitionMultimap.get() to return an empty list instead
of throwing an exception when there are no partitions associated
with a node.

Change Conductor.reconfigure() so it only connects to nodes that
have relevant partitions.

When persistence polling is *disabled*, use better logic to determine
whether a node has active partitions. The old code was in danger of
failing when alternate host name support is added, since the decision
was based on the node's hostname without considering the port.

Fix a glitch in Conductor.reconfigure() where a DCP channel would
not be removed when it stopped hosting relevant partitions.
The removal check now retains only connections to relevant nodes
instead of all nodes in the cluster.

Fix BucketConfigHelper.getNodeIndex() so it calculates the index
correctly even if some nodes don't run the data service for some reason.

ConfigHandler now logs the new bucket config (at TRACE level).

Change-Id: Id83563bc1ecdde8185243a34452815a3ff38b346
Reviewed-on: http://review.couchbase.org/104725
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: David Nault <david.nault@couchbase.com>