JDCP-110 Honor port in connection hostnames
Motivation
==========
When Couchbase runs in a container, the externally visible HTTP port
may be mapped to something other than 8091. This custom port must be
specified when bootstrapping the client.

Attempting to pass a <hostname:port> combination to
Client.Builder.hostnames() results in a connection failure because
the port is treated as part of the hostname.

As a workaround, users may build a connection string with custom ports
and pass it to Client.Builder.connectionString() instead of calling
hostnames(), but it would be nice if this weren't necessary.

Modifications
=============

Allow users to specify port when calling Client.Builder.hostnames().

Change-Id: I57bb23d1581b609ede3cbf9087d24b8d2006d002
Reviewed-on: http://review.couchbase.org/105006
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>