JDCP-119 Connection name should follow Response Time Observability RFC
Motivation
==========
The client should follow the recommendation for generating connection
names with user agent, client ID, and connection ID.

Modifications
=============
In MessageUtil.setKey call `writeBytes` to grow the buffer if necessary
instead of `setBytes` which throws an exception if the buffer is too
small.

Change MessageUtil.setKey's `key` parameter from a ByteBuf to a String
so callers don't need to repeat the conversion everywhere.

Add UserAgentBuilder so it's easy to generate user agent strings
that comply with RFC 7231.

Add ClientVersion and version.properties. Modify the POM to filter
resources so project version is available at runtime for inclusion
in user agent.

Modify DefaultConnectionNameGenerator to return the JSON specified
by the Response Time Observability RFC.

Change-Id: I48d64e8dea0e533471a92c434b1f8e962252203e
Reviewed-on: http://review.couchbase.org/105070
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>