JDCP-120 Noisy warning in log if flow control ACK happens after connection is closed
Motivation
==========
When an application acknowledges a buffer after the DCP connection is
closed, the call to writeAndFlush in ChannelFlowController.ack
fails silently. That's fine.

However, the failure also causes a WARN log entry with a scary
stack trace and a message about the rejected execution.

Modifications
=============
Skip the flow control ACK if the connection is already closed.

Change-Id: I4dcbc733eb7e2e5e7e6fb9e1ecb25949477cece4
Reviewed-on: http://review.couchbase.org/105808
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>