JDCP-121 Client reports partially-created bucket has no partitions
Motivation
==========
A bucket config's partition map may be empty for a short time
after the bucket is created. As a result, Client.numPartitions()
may return zero.

This causes problems for code that needs to know the real number
of partitions.

Modifications
=============
When awaiting the initial bucket config, ignore any configs
with missing partition info.

Change-Id: I5d141b9b05a034ea598d76b7db2be65577c16c0a
Reviewed-on: http://review.couchbase.org/105900
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>