JDCP-123 PersistencePollingHandler should not flood the logs with redundant warnings
Motivation
==========
When a connection is closed, the persistence polling handler logs one
warning for each vbucket. Each warning includes a stack trace, so for
1024 vbuckets we're talking about a lot of log lines.

This causes problems in the Elasticsearch connector integration tests,
where the server sometimes takes a while to warm up before it will
admit to owning vbuckets. The resulting log output is many megabytes
and causes the test runner to run out of memory.

Modifications
=============
Log only the first failure that triggers connection closure.

The only log message we really care about is the first one, because it
describes why the poller closed the connection. The remaining failures
are triggered by the connection closure, and don't contain any useful
information.

Change-Id: I4e70dc21ca9d1139f7767914b914b60dc81995d9
Reviewed-on: http://review.couchbase.org/108612
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>