JDCP-124 Throttle reconnection attempts
Motivation
==========
After the persistence poller closes a connection due to an error,
the client immediately reconnects. This results in a flood of failure
messages in the log until the cluster stabilizes well enough for the
poller to proceed.

Modifications
=============
Delay each reconnection attempt using exponential backoff. Reset the
delay to zero if connection is stable for 10 seconds.

Change-Id: I49153dca8b70086835aca5d7380af78b6af10ee6
Reviewed-on: http://review.couchbase.org/108623
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>