Build script: Enforce deployment preconditions
Removed the "stage" profile to streamline configuration.
Wasn't using it, and the same effect can be achieve by passing:

  -Prelease -DautoReleaseAfterClose=false

Add a "snapshot" profile that asserts the version is a snapshot
before deploying. This will make sure Jenkins never accidentally
does a real release when it only intends to deploy a snapshot.

The "release" profile will now fail a deployment if any of the
dependencies are snapshots.

Change-Id: Ib1338dfff61dab58e62a4e8e29bdbc4878d52473
Reviewed-on: http://review.couchbase.org/108805
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>