JDCP-126 Rebalance fails because DCP connection name is too long
Motivation
==========
A DCP connection whose name is near the limit of 250 bytes prevents
Couchbase Server from rebalancing. The server will be fixed eventually,
but we want a fix for customers in the short term.

Modifications
=============
Limit the user agent to 100 characters instead of 200 characters.
This means the info about the Java and OS versions may be truncated,
but that's fine.

Reorder the JSON attributes so the connection ID comes first; that way
if someone else needs to truncate the name for whatever reason,
the result is more likely to include the unique connection ID.

Change-Id: Iaada9d721b1b5ff373bdff33f3236cb5c74f14b4
Reviewed-on: http://review.couchbase.org/109212
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>