JDCP-140 Add high-level API
Provide a unified event listener. Removes guesswork about which
events are delivered to which callback handlers.

The new event listener receives objects that model the DCP events.
This makes it easier to work with the events, and prevents accidentally
using the wrong static accessor methods to interpret the contents of a
Netty ByteBuf.

Expose byte arrays to users instead of Netty ByteBufs; most (if not all)
users were converting to byte arrays anyway. This prevents potential
memory leaks caused by failure to release buffers. It also removes user
dependency on repackaged Netty so we can migrate away from SDK2 core-io.

Add an optional "safe dispatch" mode where the user's event listener
is invoked in a separate thread instead of the Netty I/O event loop.
This lets the user make blocking calls in the event handler
without disrupting the DCP client networking layer.

The "safe dispatch" mode also supports optional automatic flow control.
When enabled, the DCP client automatically issues flow control ACKs
when events are removed from the dispatch queue, prior to being handed
to the event listener. This is great for event listeners that do not
return from the callback until the event has been processed (where
"processed" can be as simple as "write to disk" or "put in a bounded
blocking queue").

Add a "HighLevelApi" example that shows how to use the new API,
complete with safe stream state persistence.

Change-Id: Ia917c35698435f980378192054eb72f2a1309850
Reviewed-on: http://review.couchbase.org/118386
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>