JDCP-137 Increase connection name limit from 148 bytes to 200 bytes
Motivation
----------
The previous limit of 148 bytes was a conservative guess. With the
resolution of MB-34280 we now know the actual safe limit is 200 bytes.

With a higher limit we can truncate the user agent string less
aggressively and retain more info about the JVM and OS.

Modifications
-------------
Increase the user agent limit from 102 bytes to 154 bytes.

Add a test that explicitly asserts the limit is enforced.

Document the 200 byte limit on the ConnectionNameGenerator interface.

Change-Id: I9efef85e86467900a123f0b1aa316ce285b75fd0
Reviewed-on: http://review.couchbase.org/119330
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>