JDCP-145 Remove dependency on core-io
Motivation
----------
Make the DCP client compatible with both SDK 3 and SDK 2. Before this
change, the DCP client could not be used with SDK 3 because of core-io
version conflicts.

Modifications
-------------
Inline the classes from core-io 1.x that the DCP client depends on,
repackaging them under `com.couchbase.client.dcp.core` so they don't
conflict with the original versions in `com.couchbase.client.core`.

Shade the Jackson and Netty dependencies under a new package,
`com.couchbase.client.dcp.deps` (again, so as not to conflict with
the versions in core-io 1.x if present). This is a breaking change
for dependent code, since the Netty ByteBuf is exposed as part of the
DCP client's low-level API. In most cases, a global search & replace on
the package name will solve that problem.

Move the examples into a separate project so the Java client they use
doesn't pollute the DCP client's class path in ways that could
introduce accidental dependencies on core-io 1.x.

Change-Id: Id3ed53edea3cd9dad8fccb490d28e645d2a3439d
Reviewed-on: http://review.couchbase.org/123941
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>