JDCP-148 Assume developer builds are latest version
Motivation
----------
Developer builds of Couchbase Server can report a version of 0.0.0.
This confuses the client, since it uses the version string to decide
which features to enabled. The client should really use capabilities
instead of version sniffing, but in the mean time treating 0.0.0 as
the latest version is a simple workaround.

Modifications
-------------
If the server reports a major version of 0, replace it with 9999.

Change-Id: I0f82dd4e4efd0c3be7fcfc9e81565e5238918c94
Reviewed-on: http://review.couchbase.org/124118
Tested-by: David Nault <david.nault@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>