JDCP-149 Store and use features from HELO response
Motivation
----------
Move away from making decisions based on version sniffing, and instead
honor the capabilities advertised by the server.

This is a groundwork for getting clustermap change notifications
via the binary protocol channel instead of HTTP.

Modifications
-------------
Add a HelloFeature enum and use it instead of raw shorts.

Parse and remember the negotiated features from the HELO response,
and use this info to decide whether to skip the "select bucket" request
(instead of deciding based on server version).

Change-Id: Ic581fc7855130adb7177cb59c73ba812079e619e
Reviewed-on: http://review.couchbase.org/124189
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>