JDCP-153 Support for collections
Motivation
----------
Client must be collections-aware in order to stream from collections
other than the default collection.

Modifications
-------------
When building the client, the user can specify the scope or collections
to stream from.

Prior to requesting a stream, the client now fetches the collections
manifest and uses it to resolve collection and scope names.

The high level API now supports all the new events related to
collections, and seamlessly handles parsing the collection ID out of
document keys.

Consumers of the low level API who wish to enable collections awareness
will need to handle the new collections control events (or ignore them,
as appropriate).

Change-Id: I5697454bb59b95601989a26599ff3994cfbb93da
Reviewed-on: http://review.couchbase.org/c/java-dcp-client/+/124861
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>