JDCP-159 Include collections manifest UID in stream offsets
Motivation
----------
Collections RFC says this is required in order to receive
consistent collections manifest updates when restarting the stream.

Modifications
-------------
Add a `collectionsManifestUid` property to StreamOffset. Use it
when opening a collections-aware stream.

A value of zero is appropriate when streaming from beginning, or if
resuming a stream that was previously not collections-aware.

Change-Id: I33d088f5267b6d22b4a8d0c79b584ba15e58c3e9
Reviewed-on: http://review.couchbase.org/c/java-dcp-client/+/125717
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>