JDCP-157 Add high-level API as a perf-runner option
Motivation
----------
Want to learn how much overhead is associated with using the
high-level API.

Modifications
-------------
Add a `high-level-api.properties` settings file that activates
the high-level API.

If that setting is enabled, register a high-level listener.
This listener doesn't have access to the original ByteBuf,
so it can't tell if the message was originally compressed.
As a result, it just counts the messages instead of collecting
compression statistics.

Use LongAdders instead of AtomicLongs, because we don't need the
additional behaviors of AtomicLongs.

Allow custom ports in connection strings.

Change-Id: I8abf12cdeb4c6e6a771827479606856a2aab40bd
Reviewed-on: http://review.couchbase.org/c/java-dcp-client/+/125735
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>