SDKQE-1946: Enable integration tests to run with internal server-build images
Motivations
Pre-release couchbase server-build images are needed for testing, which are available for Couchbase Org members.

Modifications:
In the Integration tests, the couchbase version passed as an environment variable can now specify a build number to use, ie. 6.5.0-4960. So long as the docker host is logged in to an account that is part of the Couchbase org.

Change-Id: I97235a66f295944d6323011df7e163e949102390
Reviewed-on: http://review.couchbase.org/c/java-dcp-client/+/126517
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>