JDCP-166 Bucket config parser fails if $HOST is an IPv6 literal
Motivation
----------
The bucket config JSON uses $HOST placeholders in HTTP URLs and other
contexts where it is followed by ":" and a port number. When an IPv6
literal occurs in these contexts, it must be enclosed in square
brackets, but that is not currently happening.

Modifications
-------------
Add HostAndPort.format() and formatHost() which enclose IPv6 literals
in square brackets. Use formatHost() as the $HOST replacement.

Add unit test for $HOST placeholder replacement.

Canonicalize IPv6 literals in HostAndPort. Not strictly required
for this fix, but groundwork for relying more heavily on HostAndPort
for node equality.

Change-Id: Idbfe336a391bedb6fc3ad0feeda3dc6967f3c1a3
Reviewed-on: http://review.couchbase.org/c/java-dcp-client/+/127331
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>