Gardening: Rename NetworkResolution.custom to valueOf
and reduce a tiny bit of code duplication in DcpBucketConfig.

Motivation
----------
The "custom" netowrk resolution might not actually be custom...
we might in fact be translating from the Java SDK's version to the
DCP client's version.

This change also brings the DCP client in line with SDK 3, which also
uses `valueOf` instead of `custom`

Change-Id: Ie5dfb68e8b4d6beb7e85aef9dec4ccfd6d0efd28
Reviewed-on: http://review.couchbase.org/c/java-dcp-client/+/128403
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>