Gardening: Ensure initial configuration has valid partition count
Motivation
----------
Address a regression in connectWaitsForPartitionMap() integration test,
which asserts that the partition count is valid.

This primarily affects integration tests which create a bucket
and then immediately start streaming from it.

Modifications
-------------
Move the "skip config with invalid partition count" logic from
Conductor.connect() to the config stream handler which applies the config.
This prevents the bogus config from being applied and made visible to
user code.

Change-Id: I4a3c295396796368dade4152e654a0301a639ab8
Reviewed-on: http://review.couchbase.org/c/java-dcp-client/+/128790
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>