Gardening: Reduce code duplication between builders
Motivation
----------
ClientEnvironment.Builder used to duplicate most of the
properties of Client.Builder, making it tedious to
add a new config property.

Modifications
-------------
Move ClientEnvironment to be an inner class of Client
so it can access the properties of Client.Builder.

Use the same Client.Builder to create both the Client
and the Client.Environment.

Change-Id: I0d1220f075566e60f030d4d06788bc2a5380f070
Reviewed-on: http://review.couchbase.org/c/java-dcp-client/+/128799
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>