JDCP-169 Support streaming documents without contents
Motivation
----------
Anticipating a use case where the DCP stream is used
for cache invalidation, or some other scenario where the
document contents are not interesting.

Modifications
-------------
Add a `noValue()` method to the client builder. When this method
is called, the client passes OpenConnectionFlag.NO_VALUE
when opening DCP connections.

Remove the DcpOpenConnectionRequest.Type enum since it duplicates
the functionality of OpenConnectionFlag.

Remove unused Client.Environment.builder() method
(missed this in the prior change that made Client and
Client.Environment share the same builder).

Change-Id: Iba55f17a8692b394f879d3d4ddef2c53813eaf0b
Reviewed-on: http://review.couchbase.org/c/java-dcp-client/+/128809
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>