Gardening: Remove unnecessary partition state assignments
Motivation
----------
In some places where the partition state is modified, the call to
SessionState.get(partition) is followed by SessionState.set().

The call to `set` has no effect, since SessionState returns the
PartitionState it owns; it doesn't make a defensive copy. It's not clear
why this call is there -- maybe from a historical revision where
SessionState returned a copy?

Modifications
-------------
Remove the unnecessary calls to SessionState.set.

Consolidate the handled of Mutation, Deletion, and Expiry messages
in the dataEventHandler decorator. (Regardless of message type, all the
decorator does is set the partition state's seqno to the message seqno.)

Change-Id: Ic5913783f542b08152fe3e043401506f6d12d56f
Reviewed-on: http://review.couchbase.org/c/java-dcp-client/+/137166
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>